Make elm-analyse json injectable
