+++ src/providers/diagnostics/elmLsDiagnostics.ts
@@ -24,46 +24,13 @@ import { Utils } from "../../util/utils";
 import { IDiagnostic } from "./diagnosticsProvider";
 import * as path from "path";
 import { SyntaxNodeMap } from "../../compiler/utils/syntaxNodeMap";
+import { IElmAnalyseJsonService } from "./elmAnalyseJsonService";
 
-interface IElmAnalyseJson {
-  checks?: {
-    BooleanCase?: boolean;
-    DebugLog?: boolean; // We don't support this
-    DebugTodo?: boolean; // We don't support this
-    DropConcatOfLists?: boolean;
-    DropConsOfItemAndList?: boolean;
-    DuplicateImport?: boolean; // We don't support this as elm-format will fix this
-    DuplicateImportedVariable?: boolean; // We don't support this as elm-format will fix this
-    ExposeAll?: boolean; // We don't support this
-    FileLoadFailed?: boolean; // We don't support this as it makes no sense for us
-    NoUncurriedPrefix?: boolean;
-    FunctionInLet?: boolean; // We don't support this
-    ImportAll?: boolean; // We don't support this
-    MapNothingToNothing?: boolean;
-    MultiLineRecordFormatting?: boolean; // We don't support this
-    NoTopLevelSignature?: boolean; // We don't support this as we get it via type inference already
-    SingleFieldRecord?: boolean;
-    TriggerWords?: string[]; // We don't support this
-    UnnecessaryListConcat?: boolean;
-    UnnecessaryParens?: boolean; // We don't support this as elm-format will fix these anyway
-    UnnecessaryPortModule?: boolean;
-    UnusedImport?: boolean;
-    UnusedImportAlias?: boolean;
-    UnusedImportedVariable?: boolean;
-    UnusedPatternVariable?: boolean;
-    UnusedTopLevel?: boolean;
-    UnusedTypeAlias?: boolean;
-    UnusedValueConstructor?: boolean;
-    UnusedVariable?: boolean;
-    UseConsOverConcat?: boolean;
-  };
-  excludedPaths?: string[];
-}
 export class ElmLsDiagnostics {
   private language: Language;
   private elmWorkspaceMatcher: ElmWorkspaceMatcher<URI>;
   private connection: Connection;
-  private elmAnalyseJson = new Map<string, IElmAnalyseJson>();
+  private elmAnalyseJsonService: IElmAnalyseJsonService;
 
   private readonly exposedValuesAndTypesQuery: Query;
   private readonly exposedValueAndTypeUsagesQuery: Query;
@@ -91,6 +58,9 @@ export class ElmLsDiagnostics {
     this.language = container.resolve<Parser>("Parser").getLanguage();
     this.elmWorkspaceMatcher = new ElmWorkspaceMatcher((uri) => uri);
     this.connection = container.resolve<Connection>("Connection");
+    this.elmAnalyseJsonService = container.resolve<IElmAnalyseJsonService>(
+      "ElmAnalyseJsonService",
+    );
 
     this.exposedValuesAndTypesQuery = this.language.query(
       `
@@ -422,7 +392,9 @@ export class ElmLsDiagnostics {
     sourceFile: ISourceFile,
     program: IProgram,
   ): IDiagnostic[] => {
-    const elmAnalyseJson = this.getElmAnalyseJson(program.getRootPath().fsPath);
+    const elmAnalyseJson = this.elmAnalyseJsonService.getElmAnalyseJson(
+      program.getRootPath().fsPath,
+    );
     const tree = sourceFile.tree;
     const uri = sourceFile.uri;
     const rootPath = program.getRootPath().fsPath;
@@ -496,29 +468,6 @@ export class ElmLsDiagnostics {
     }
     return [];
   };
-
-  private getElmAnalyseJson(workspacePath: string): IElmAnalyseJson {
-    const cached = this.elmAnalyseJson.get(workspacePath);
-
-    if (cached) {
-      return cached;
-    }
-
-    let elmAnalyseJson = {};
-    try {
-      elmAnalyseJson = require(path.join(
-        workspacePath,
-        "elm-analyse.json",
-      )) as IElmAnalyseJson;
-    } catch {
-      this.connection.console.info(
-        "No elm-analyse.json found, enabling all diagnostic checks.",
-      );
-    }
-
-    this.elmAnalyseJson.set(workspacePath, elmAnalyseJson);
-    return elmAnalyseJson;
-  }
 
   private getUnusedImportDiagnostics(tree: Tree): IDiagnostic[] {
     const diagnostics: IDiagnostic[] = [];
